requirements.txt
phonemizer==3.2.1
munch
accelerate
pydub
nltk
g2p_en
num2words
inflect
unidecode
pyyaml
librosa
scipy
matplotlib
soundfile
torch
torchaudio
transformers
einops
einops-exts
tqdm
omegaconf
huggingface_hub
